move gemsSchema to a separate file under the folder/schema
