+++ src/hooks/useGems.ts
@@ -2,22 +2,7 @@ import { useQuery } from "@tanstack/react-query";
 import axios from "axios";
 import Gem from "../types/gem";
 import { sortGemsByDate } from "../utils/helpers";
-import { z } from "zod";
-
-const gemsSchema = z.array(
-  z.object({
-    id: z.string(),
-    title: z.string(),
-    description: z.string(),
-    image: z.string(),
-    date: z.object({
-      seconds: z.number(),
-      nanoseconds: z.number(),
-    }),
-    categoryId: z.string(),
-    gemmerId: z.string(),
-  })
-);
+import gemsSchema from "../schema/gems.schema";
 
 const useGems = () => {
   return useQuery(
